requirements.txt
python-dotenv==0.19.2
boto3==1.20.15
zipfile36==0.1.3
ftputil==5.0.1
git+https://github.com/SCOAP3/hepcrawl.git@python3-compatible